[Frontend] Login form relocated to home page
