+++ client/src/app/components/home/home.component.html
@@ -1,20 +1,45 @@
-<div class="container mt-5">
-    <div style="text-align: center" *ngIf="!registerMode">
+<div class="container mt-5 d-flex">
+    <div class="text-center mt-5 mr-5">
         <h1>Find your match</h1>
         <p class="lead">Let's go and view your matches.. all you need to do is to sign up!</p>
-        <div class="text-center">
-            <button id="registerBtn" class="btn btn-primary btn-lg mr-2" (click)="registerToggle()">Register</button>
-            <button class="btn btn-info btn-lg">Learn more</button>
-            <button class="btn btn-success btn-lg ml-2" type="button" (click)="signInGoogle()">
+        <button class="btn btn-info btn-lg">Learn more</button>
+    </div>
+    <div class="container text-center" *ngIf="!registerMode">
+        <h2>Login</h2>
+        <form id='loginform' #loginForm="ngForm" class="ngForm mt-2" (ngSubmit)="login()" autocomplete="off">
+            <div class="form-group">
+                <input
+                    name="username"
+                    [(ngModel)]="creds.username" 
+                    class="form-control mr-sm-2" 
+                    type="text" 
+                    placeholder="Username">
+            </div>
+            <div class="form-group">
+                <input 
+                    name="password"
+                    [(ngModel)]="creds.password"
+                    class="form-control mr-sm-2"
+                    type="password"
+                    placeholder="Password">
+            </div>
+            <button class="btn btn-outline-primary btn-block" type="submit">Login</button>
+        </form>
+        <div class="register">
+            Do not have an account? 
+            <a href="#" id="registerBtn" (click)="registerToggle()">Register</a>
+        </div>
+        <div class="mt-3">
+            <button class="btn btn-outline-success btn-lg" type="button" (click)="signInGoogle()">
                 <i class="fa-google fa"></i>
-                Google</button>
-            <button class="btn btn-info btn-lg" (click)="signOut()">Log out</button>
+                Sign in with Google
+            </button>
         </div>
     </div>
 
     <div *ngIf="registerMode" class="container">
         <div class="row justify-content-center">
-            <div class="col-4">
+            <div class="col-8">
                 <app-register (cancelRegister)="cancelRegisterMode($event)"></app-register>
             </div>
         </div>
