+++ client/src/app/components/home/home.component.spec.ts
@@ -1,16 +1,30 @@
+import { HttpClientModule } from '@angular/common/http';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
+import { Router } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
+import { ToastrModule } from 'ngx-toastr';
+import { from } from 'rxjs';
+import { AccountService } from 'src/app/services/account.service';
 
 import { HomeComponent } from './home.component';
 
 describe('HomeComponent', () => {
   let component: HomeComponent;
   let fixture: ComponentFixture<HomeComponent>;
+  let accService: AccountService;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [ HomeComponent ],
+      imports: [
+        HttpClientModule,
+        ToastrModule.forRoot(),
+        FormsModule,
+        RouterTestingModule
+      ],
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
     })
     .compileComponents();
@@ -19,6 +33,7 @@ describe('HomeComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(HomeComponent);
     component = fixture.componentInstance;
+    accService = fixture.debugElement.injector.get(AccountService);
     fixture.detectChanges();
   });
 
@@ -43,4 +58,30 @@ describe('HomeComponent', () => {
     let regComponent = fixture.debugElement.nativeElement.querySelector('app-register');
     expect(regComponent).toBeNull();
   });
+
+
+  describe('login', () => {
+    it('should call the server to login', () => {
+      let creds = { username: 'user', password: 'pass'};
+      component.creds = creds;
+      let loginSpy = spyOn(accService, 'login').and.returnValue(from([]));
+  
+      component.login();
+  
+      expect(loginSpy).toHaveBeenCalledWith(creds);
+    });
+    
+    // TODO: Not working properly
+    xit('should redirect to members if login is successful', () => {
+      let router = fixture.debugElement.injector.get(Router);
+      let routeSpy = spyOn(router, 'navigateByUrl');
+      let loginSpy = spyOn(accService, 'login').and.returnValue(from([]));
+  
+      component.login();
+
+      loginSpy.calls.mostRecent().returnValue.subscribe(() => {
+        expect(routeSpy).toHaveBeenCalledWith('/members');
+      });
+    });
+  })
 });
