+++ client/src/app/components/navbar/navbar.component.html
@@ -28,22 +28,6 @@
                     <a href="" class="dropdown-item" (click)="logout()">Logout</a>
                 </div>
             </div>
-
-            <form id='loginform'#loginForm="ngForm" class="form-inline mt-2 mt-md-0" (ngSubmit)="login()" autocomplete="off" *ngIf="(accountService.currentUser$ | async) === null">
-                <input
-                    name="username"
-                    [(ngModel)]="creds.username" 
-                    class="form-control mr-sm-2" 
-                    type="text" 
-                    placeholder="Username">
-                <input 
-                    name="password"
-                    [(ngModel)]="creds.password"
-                    class="form-control mr-sm-2"
-                    type="password"
-                    placeholder="Password">
-                <button class="btn btn-outline-success" type="submit">Login</button>
-            </form>
         </div>
     </div>
 </nav>
