+++ client/src/app/components/navbar/navbar.component.spec.ts
@@ -70,11 +70,6 @@ describe('NavbarComponent', () => {
       expect(userMenu).toBeTruthy();
     })
     
-    it('should NOT render login form', () => {
-      let loginform: HTMLElement = fixture.debugElement.query(By.css('#loginform')) as any;  
-      expect(loginform).toBeFalsy();
-    })
-    
     it("should render user's username or knownas property", () => {
       let name: HTMLElement = fixture.debugElement.query(By.css('.dropdown-toggle')).nativeElement;
       expect(name.textContent.toLowerCase()).toContain(user.username.toLowerCase() || user.knownAs.toLowerCase());
@@ -93,37 +88,6 @@ describe('NavbarComponent', () => {
       let userMenu: HTMLElement = fixture.debugElement.query(By.css('#userMenu')) as any;
       expect(userMenu).toBeFalsy();
     })
-    
-    it('should render login form', () => {
-      let loginform: HTMLElement = fixture.debugElement.query(By.css('#loginform')) as any;  
-      expect(loginform).toBeTruthy();
-    })
-  })
-
-  
-  describe('login', () => {
-    it('should call the server to login', () => {
-      let creds = { username: 'user', password: 'pass'};
-      component.creds = creds;
-      let loginSpy = spyOn(accService, 'login').and.returnValue(from([]));
-  
-      component.login();
-  
-      expect(loginSpy).toHaveBeenCalledWith(creds);
-    });
-    
-    // TODO: Not working properly
-    xit('should redirect to members if login is successful', () => {
-      let router = fixture.debugElement.injector.get(Router);
-      let routeSpy = spyOn(router, 'navigateByUrl');
-      let loginSpy = spyOn(accService, 'login').and.returnValue(from([]));
-  
-      component.login();
-
-      loginSpy.calls.mostRecent().returnValue.subscribe(() => {
-        expect(routeSpy).toHaveBeenCalledWith('/members');
-      });
-    });
   })
